[refactor] 카테고리 불러와서 적용하기
